fix(flotilla): send fractional CPU to Ray autoscaler and reject non-finite min_cpu_per_task
Two follow-ups from review of the min_cpu_per_task wiring:

- The Ray autoscaler bundle still rounded CPU up via `num_cpus().ceil()`,
  so a configured `min_cpu_per_task=0.1` reached `request_resources` as
  `{"CPU": 1}` (issue #7123). Add `TaskResourceRequest::autoscale_bundle`
  which keeps CPU fractional (GPU/memory stay integer, zero values dropped)
  and build the Ray bundle as a `PyDict` with a float CPU. Update the
  `try_autoscale` type hint accordingly.

- `min_cpu_per_task` validation let NaN/inf through (`nan <= 0` and
  `inf <= 0` are both false). Use `is_finite() && > 0` on both the Python
  setter and the env path.

Add Rust unit tests for the fractional bundle and env validation, and a
Python test covering acceptance of fractional values and rejection of
0/-0.5/nan/inf.

+++ daft/runners/flotilla.py
@@ -566,7 +566,7 @@ def start_ray_workers(
     return handles
 
 
-def try_autoscale(bundles: list[dict[str, int]]) -> None:
+def try_autoscale(bundles: list[dict[str, float]]) -> None:
     from ray.autoscaler.sdk import request_resources
 
     request_resources(
